refactor: create base query class
